control-plane/requirements.txt
# Core
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.9.2
python-dotenv==1.0.1

# Database
SQLAlchemy==2.0.25
asyncpg==0.29.0
aiosqlite==0.20.0
alembic==1.13.1

# HTTP & Async
httpx==0.25.2
aiofiles==23.2.1

# Security & Auth
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
PyJWT==2.10.1

# Production
gunicorn==21.2.0
python-multipart==0.0.6

# Monitoring & Logging
python-json-logger==2.0.7
prometheus_client==0.19.0

# Testing (dev dependencies would go in requirements-dev.txt)
pytest==7.4.3
pytest-asyncio==0.21.1
